ENG-7529 seperate retry logic from event sender - seperate retry logic from event sender so it can be reused in the remote config getter. - add generic response type to retry logic
