+++ NeuroID/src/main/java/com/neuroid/tracker/service/NIDApiService.kt
@@ -1,9 +1,11 @@
 package com.neuroid.tracker.service
 
+import com.neuroid.tracker.models.NIDRemoteConfig
 import okhttp3.RequestBody
 import okhttp3.ResponseBody
 import retrofit2.Call
 import retrofit2.http.Body
+import retrofit2.http.GET
 import retrofit2.http.POST
 import retrofit2.http.Path
 
@@ -13,4 +15,9 @@ interface NIDApiService {
         @Body requestBody: RequestBody,
         @Path("key") key: String,
     ): Call<ResponseBody>
+
+    @GET("/mobile/{key}.json")
+    fun getConfig(
+        @Path("key") key: String
+    ): Call<NIDRemoteConfig>
 }
